+++ app/src/main/java/com/mck/weathernow/model/Period.java
@@ -4,7 +4,7 @@ package com.mck.weathernow.model;
  * Created by Michael on 5/28/2016.
  */
 public class Period {
-    public Integer dt;
+    public Long dt;
     public Main main;
     public Weather[] weather;
     public Clouds clouds;
